feature/evetual-conectivity-create-alert (#37)
* feat: implement draft alerts and offline support (eventual conectivity for alerts)

This commit introduces draft alert functionality and offline support for alert creation. It adds a new `DraftAlertEntity` for local storage of draft alerts, a `DraftAlertDao` for database interactions, and a `DraftAlertRepository` for data handling. Additionally, a `ConnectivityMonitor` is implemented to observe network status changes. The `AlertRepository` is modified to handle the creation of alerts, now returning an `AlertDomain` object. A notification service (`AlertNotificationService`) is added to alert users about available draft alerts. The `DraftAlertsScreen` is introduced to display and manage draft alerts. The AlertCreateScreen has been updated to support saving draft alerts when offline and to load a previous draft. The app database is updated to version 2 to include `DraftAlertEntity`.

* feat: implement draft alerts and offline support (eventual conectivity for alerts)

This commit introduces draft alert functionality and offline support for alert creation. It adds a new `DraftAlertEntity` for local storage of draft alerts, a `DraftAlertDao` for database interactions, and a `DraftAlertRepository` for data handling. Additionally, a `ConnectivityMonitor` is implemented to observe network status changes. The `AlertRepository` is modified to handle the creation of alerts, now returning an `AlertDomain` object. A notification service (`AlertNotificationService`) is added to alert users about available draft alerts. The `DraftAlertsScreen` is introduced to display and manage draft alerts. The AlertCreateScreen has been updated to support saving draft alerts when offline and to load a previous draft. The app database is updated to version 2 to include `DraftAlertEntity`.